Added docker container support for Linux
My Ubuntu test stack has been upgraded to 20.04 and some packages required for Dolphin aren't supported yet for this release. To counteract this, I created a dockerfile to test if the Linux environment works. If you are modifying a local release, ensure you modify the install.sh and Dockerfile accordingly (Use ADD for Ishiiruka)