MAJOR CODE PUSH
Added receiver optimizations that provides near lagless performance and communication between engine and the child processes. Added controller canvas but it still needs to be optimized heavily, the Queue might be removed in favor of list if it proves faster. The canvas itself is unoptimized and definitely proves to be a barrier on displaying the data in a quick way for the user.